Battle Scene Updates
Chat Screen now displays each team and then allows us to manually move to the next turn of the battle with a button. Results are shown at the bottom of the screen.

With these now incremented and the JSON being properly parsed, we can begin animating each step in FOR LOOP and show the results after the LOOP closes. Need to package the rewards into the JSON to also display with the results.