Cleaner Battle Sequence
Updated disposals, timing, and sequencing so that animations play in the right order without overlap. Animations now stop before the end of the battle and dispose step to avoid errors.

Need to work on the opening of the battle so elements don't error out waiting for data that isn't there yet. More to come.