Major Battle Scene Improvements
No more Flutter errors, background platforms now incorporated, damage/move Containers are now separate and switch transparency based on the state of the reverseAttackAnimationController.

All items now live in the stack so monster images can move offscreen using the Positioned widget. This will need to be updated based on screen size but works for a first pass. More to come.