New Squad Position Pages & Pop Menu for Positions
Squad 1/2/3 pages created for Spirit List page with live data from Firebase.

Pop menu (dummy) created so a player can change the active members of a squadron by clicking a button and selecting 1/2/3. Need to update to be accurate and to only allow one ID in each slot at a time.