New Battle Page WIP
Battle navigation now leads to a WIP battle hub. Contains the user's squad and a list of users to battle. Need to update so you're actually battling the person you click on. Need to brainstorm a better way to present the list of users so new players don't have to scroll forever to find a properly matched opponent.