WIP Battle Sequence - Firebase Connection
Working with a new battle sequence NodeJS file to connect to Firebase to pull in the teams. Might need to rework data structure in the cloud to pull data from subcollections of 'users' collection to be more efficient (across the whole app).

Stopped at the point I could pull all documentation for the 'users' collection in the console.log(). Need to continue working on this to pull actual monster data for each team to fill the battle function criteria.