Battle Function Calls Firebase Squadrons
All 6 spirits for the 2 teams are pulled from Firebase. The user ID's for attacker/defender are still hard-coded and will need a way to connect via the app.

Battle Scene code in Flutter needs to be updated with the JSON output wherever it lives rather than the hard-coded data that's currently set for now.