Earth Type files & Spirit List View
Included proper file names for the earth type. Updated the Spirit List view to point to the proper place in Firebase but need to get a properly uniform format for spirit data to be stored so I can easily reuse code/assets. Also updated to a new version of Flutter if any files break.